Better german translation
Right now it's "ein Uhr", which works if you actually mean one o'clock, but not when it's five past one, then in German it needs to be "eins", i.e "fünf nach eins". This way, one o'clock would still be wrong, because it would be "eins uhr" instead of "ein uhr", but this way it would be wrong ten minutes each day instead of 2x55 minutes.

"fünf vor halb" and "fünf nach halb" sound more natural in German and should also avoid the other current problem, that fünfund & zwanzig are both one letter to long so they are shown as "fünfun zwanzi", which looks weird. This is still a problem with zwanzi, but that could only be changed through a smaller font.